Makes cgo flags via environment superfluous
Adds the "-static" flag to the LDFLAGS via the `yara_static` build tag.

The library info about openssl seems to be picked up automatically
because I built libyara with static flags.
If there is ever complaint about openssl not being picked up I might
need to add it to the build tags.